fix(api): include progress-steps methods
