+++ src/instanceMethods.js
@@ -2,4 +2,5 @@ export * from './instanceMethods/hideLoading'
 export * from './instanceMethods/getInput'
 export * from './instanceMethods/enable-disable-elements'
 export * from './instanceMethods/show-reset-validation-error'
+export * from './instanceMethods/progress-steps'
 export * from './instanceMethods/_main'
