+++ src/utils/dom/index.js
@@ -4,3 +4,4 @@ export * from './getters'
 export * from './parseHtmlToContainer'
 export * from './animationEndEvent'
 export * from './measureScrollbar'
+export * from './renderers/renderProgressSteps'
