+++ src/utils/dom/renderers/renderProgressSteps.js
@@ -1,0 +1,38 @@
+import { swalClasses } from '../../classes.js'
+import { warn } from '../../utils.js'
+import * as dom from '../../dom/index'
+import sweetAlert from '../../../sweetalert2'
+
+export const renderProgressSteps = (params) => {
+  let progressStepsContainer = dom.getProgressSteps()
+  let currentProgressStep = parseInt(params.currentProgressStep === null ? sweetAlert.getQueueStep() : params.currentProgressStep, 10)
+  if (params.progressSteps && params.progressSteps.length) {
+    dom.show(progressStepsContainer)
+    dom.empty(progressStepsContainer)
+    if (currentProgressStep >= params.progressSteps.length) {
+      warn(
+        'Invalid currentProgressStep parameter, it should be less than progressSteps.length ' +
+        '(currentProgressStep like JS arrays starts from 0)'
+      )
+    }
+    params.progressSteps.forEach((step, index) => {
+      let circle = document.createElement('li')
+      dom.addClass(circle, swalClasses.progresscircle)
+      circle.innerHTML = step
+      if (index === currentProgressStep) {
+        dom.addClass(circle, swalClasses.activeprogressstep)
+      }
+      progressStepsContainer.appendChild(circle)
+      if (index !== params.progressSteps.length - 1) {
+        let line = document.createElement('li')
+        dom.addClass(line, swalClasses.progressline)
+        if (params.progressStepsDistance) {
+          line.style.width = params.progressStepsDistance
+        }
+        progressStepsContainer.appendChild(line)
+      }
+    })
+  } else {
+    dom.hide(progressStepsContainer)
+  }
+}
